utilizando for, while, e ifs

+++ sintaxe-variaveis-e-fluxo/src/Fatorial.java
@@ -3,9 +3,9 @@ public class Fatorial {
 
 	public static void main(String[] args) {
 
-		int fatorial = 1;
+		long fatorial = 1;
 
-		for (int i = 1; i < 11; i++) {
+		for (long i = 1; i < 11; i++) {
 			fatorial *= i;
 			System.out.println("fatorial de " + i + "=" + fatorial);
 		}
